capability-tests/pyproject.toml
[tool.poetry]
name = "wayland-capability-tests"
version = "0.1.0"
description = ""
authors = ["Stefan Schneider-Kennedy <code@stefansk.name>"]
readme = "README.md"
package-mode = true

[tool.poetry.dependencies]
python = "^3.13"
python-wayland = "^0.6.0"
typer = "^0.15.2"

[tool.poetry.group.dev.dependencies]
lxml = "^5.3.1"
requests = "^2.32.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
wayland-capability-tests = "wayland_capability_tests.app:app"
